Specify versions for some packages to work around a zypper bug.
SUSE's "zypper install" is case-insensitive on package names:
https://bugzilla.novell.com/show_bug.cgi?id=749784

The problem is triggered by packages that got renamed and differ
only in letter case.

The workaround is to specify a version requirement that matches
only one of the conflicting packages.

FIXME the packages come from non-default repos, figure out
where to add those:

    # for euca2ools only
    zypper ar http://download.opensuse.org/repositories/Virtualization:/Cloud:/OpenStack:/SUSE/SLE_11_SP2 vcos
    # various others
    zypper ar http://download.opensuse.org/repositories/devel:/languages:/python/SLE_11_SP2 python
    # user has to confirm the repo keys
    zypper ref